# crates/twerk-web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Twerk API",
    "description": "Task scheduling and job queue management API",
    "contact": {
      "name": "Twerk Contributors"
    },
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/scheduled-jobs": {
      "get": {
        "tags": [
          "super::handlers::scheduled::read"
        ],
        "summary": "GET /scheduled-jobs",
        "operationId": "list_scheduled_jobs_handler",
        "responses": {
          "200": {
            "description": "List of scheduled jobs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ScheduledJobSummary"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "super::handlers::scheduled::create"
        ],
        "summary": "POST /scheduled-jobs",
        "operationId": "create_scheduled_job_handler",
        "requestBody": {
          "description": "Scheduled job definition as JSON or YAML",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduledJobBody"
              }
            },
            "application/x-yaml": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduledJobBody"
              }
            },
            "application/yaml": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduledJobBody"
              }
            },
            "text/yaml": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduledJobBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Scheduled job created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduledJobSummary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid scheduled job payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/scheduled-jobs/{id}": {
      "get": {
        "tags": [
          "super::handlers::scheduled::read"
        ],
        "operationId": "get_scheduled_job_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Scheduled job ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scheduled job found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduledJob"
                }
              }
            }
          },
          "404": {
            "description": "Scheduled job not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "super::handlers::scheduled::lifecycle"
        ],
        "operationId": "delete_scheduled_job_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Scheduled job ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scheduled job deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "Scheduled job not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/scheduled-jobs/{id}/pause": {
      "put": {
        "tags": [
          "super::handlers::scheduled::lifecycle"
        ],
        "operationId": "pause_scheduled_job_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Scheduled job ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scheduled job paused",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Scheduled job cannot be paused",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "404": {
            "description": "Scheduled job not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/scheduled-jobs/{id}/resume": {
      "put": {
        "tags": [
          "super::handlers::scheduled::lifecycle"
        ],
        "operationId": "resume_scheduled_job_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Scheduled job ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scheduled job resumed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Scheduled job cannot be resumed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "404": {
            "description": "Scheduled job not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "tags": [
          "super::handlers::tasks"
        ],
        "summary": "GET /tasks/{id}\n# Errors",
        "operationId": "get_task_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{id}/log": {
      "get": {
        "tags": [
          "super::handlers::tasks"
        ],
        "summary": "GET /tasks/{id}/log\n# Errors",
        "operationId": "get_task_log_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Page size (max 100)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "Search query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated task log entries",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_TaskLogPart"
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AutoDelete": {
        "type": "object",
        "description": "`AutoDelete` defines automatic cleanup configuration.",
        "properties": {
          "after": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "CreateScheduledJobBody": {
        "type": "object",
        "properties": {
          "auto_delete": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AutoDelete"
              }
            ]
          },
          "cron": {
            "type": [
              "string",
              "null"
            ]
          },
          "defaults": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JobDefaults"
              }
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "inputs": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "output": {
            "type": [
              "string",
              "null"
            ]
          },
          "permissions": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          },
          "secrets": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "tags": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "tasks": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/Task"
            }
          },
          "webhooks": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/Webhook"
            }
          }
        }
      },
...
      "JobDefaults": {
        "type": "object",
        "description": "`JobDefaults` defines default values for job tasks.",
        "properties": {
          "limits": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TaskLimits"
              }
            ]
          },
          "priority": {
            "type": "integer",
            "format": "int64"
          },
          "queue": {
            "type": [
              "string",
              "null"
            ]
          },
          "retry": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TaskRetry"
              }
            ]
          },
          "timeout": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "MessageResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "Page_ScheduledJobSummary": {
        "type": "object",
        "description": "Page represents a paginated result set",
        "required": [
          "items",
          "number",
          "size",
          "total_pages",
          "total_items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "`ScheduledJobSummary` provides a summary view of a scheduled job.",
              "properties": {
                "createdAt": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "date-time"
                },
                "createdBy": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/User"
                    }
                  ]
                },
                "cron": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "description": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "id": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/string"
                    }
                  ]
                },
                "inputs": {
                  "type": [
                    "object",
                    "null"
                  ],
                  "additionalProperties": {
                    "type": "string"
                  },
                  "propertyNames": {
                    "type": "string"
                  }
                },
                "name": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "state": {
                  "$ref": "#/components/schemas/ScheduledJobState"
                },
                "tags": {
                  "type": [
                    "array",
                    "null"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "description": "Items in the current page"
          },
          "number": {
            "type": "integer",
            "format": "int64",
            "description": "Current page number"
          },
          "size": {
            "type": "integer",
            "format": "int64",
            "description": "Page size"
          },
          "total_items": {
            "type": "integer",
            "format": "int64",
            "description": "Total number of items"
          },
          "total_pages": {
            "type": "integer",
            "format": "int64",
            "description": "Total number of pages"
          }
        }
      },
      "Page_TaskLogPart": {
        "type": "object",
        "description": "Page represents a paginated result set",
        "required": [
          "items",
          "number",
          "size",
          "total_pages",
          "total_items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "`TaskLogPart` represents a part of a task's log output.",
              "properties": {
                "contents": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "createdAt": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "date-time"
                },
                "id": {
                  "type": "string"
                },
                "number": {
                  "type": "integer",
                  "format": "int64"
                },
                "taskId": {
                  "type": "string"
                }
              }
            },
            "description": "Items in the current page"
          },
          "number": {
            "type": "integer",
            "format": "int64",
            "description": "Current page number"
          },
          "size": {
            "type": "integer",
            "format": "int64",
            "description": "Page size"
          },
          "total_items": {
            "type": "integer",
            "format": "int64",
            "description": "Total number of items"
          },
          "total_pages": {
            "type": "integer",
            "format": "int64",
            "description": "Total number of pages"
          }
        }
      },
...
      "ScheduledJob": {
        "type": "object",
        "description": "`ScheduledJob` represents a scheduled job.",
        "properties": {
          "autoDelete": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AutoDelete"
              }
            ]
          },
          "createdAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "createdBy": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "cron": {
            "type": [
              "string",
              "null"
            ]
          },
          "defaults": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JobDefaults"
              }
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/string"
              }
            ]
          },
          "inputs": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "output": {
            "type": [
              "string",
              "null"
            ]
          },
          "permissions": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          },
          "secrets": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "state": {
            "$ref": "#/components/schemas/ScheduledJobState"
          },
          "tags": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "tasks": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/Task"
            }
          },
          "webhooks": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/Webhook"
            }
          }
        }
      },
      "ScheduledJobState": {
        "type": "string",
        "description": "State for a [`ScheduledJob`].",
        "enum": [
          "ACTIVE",
          "PAUSED"
        ]
      },
      "ScheduledJobSummary": {
        "type": "object",
        "description": "`ScheduledJobSummary` provides a summary view of a scheduled job.",
        "properties": {
          "createdAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "createdBy": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "cron": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/string"
              }
            ]
          },
          "inputs": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "state": {
            "$ref": "#/components/schemas/ScheduledJobState"
          },
          "tags": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
      "StatusResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "disabled": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "username": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "string": {
        "type": "string"
      }
    }
  }
}